feat: create src/modules layout and move core files
- Create module directories: bot, ai, trivia, freegames, whosplaying, infra, events
- Move core files into respective modules maintaining directory structure
- Update all index.ts files to use local imports within modules
- Update cross-module imports to use module public APIs
- Update test imports to new module structure
- Update serverless.yml to point to new function location
- Preserve all existing functionality and behavior

Closes #10
