+++ src/index.ts
@@ -1,6 +1,6 @@
 import 'dotenv/config'
 
-import createBot from './bot.js'
+import { createBot } from './modules/bot/index.js'
 
 const bot = createBot(process.env.BOT_TOKEN ?? '')
 bot.launch()
