+++ src/modules/ai/index.ts
@@ -1,4 +1,4 @@
-export { generate } from '../../ai/engine.js'
-export { text } from '../../ai/output.js'
-export { whosplayingHistory } from '../../ai/history.js'
-export { triviaExpert, whosplayingExpert } from '../../ai/system.prompt.js'
+export { generate } from './ai/engine.js'
+export { text } from './ai/output.js'
+export { whosplayingHistory } from './ai/history.js'
+export { triviaExpert, whosplayingExpert } from './ai/system.prompt.js'
