+++ src/modules/events/index.ts
@@ -1,1 +1,1 @@
-export * from '../../events.js'
+export * from './events.js'
