+++ src/modules/freegames/freegames/freegames.interface.ts
