+++ src/modules/infra/mocks/ai.mock.ts
