+++ src/modules/trivia/index.ts
@@ -1,2 +1,3 @@
-export * from '../../trivia/trivia.helper.js'
-export * from '../../trivia/trivia.service.js'
+export * from './trivia/trivia.helper.js'
+export * from './trivia/trivia.service.js'
+export * from './trivia/trivia.interface.js'
