+++ src/modules/trivia/trivia/trivia.helper.ts
