+++ src/modules/trivia/trivia/trivia.interface.ts
