+++ src/modules/trivia/trivia/trivia.service.ts
