+++ src/modules/whosplaying/index.ts
@@ -1,2 +1,3 @@
-export * from '../../whosplaying/guild.helper.js'
-export * from '../../whosplaying/guild.service.js'
+export * from './whosplaying/guild.helper.js'
+export * from './whosplaying/guild.service.js'
+export * from './whosplaying/guild.interface.js'
